sql: move analyzeExpr to a 'p_' file where it belongs
